Progress on computation of routing stats and some minor changes
* delegated_stats_v3.py updated for the ‘-y’ (year) option not to be
mandatory. If a year is not provided, statistics for all the available
years are computed.
* DelegatedHandler.py updated to accept null year and to consider the
cases in which a region is not available for a country code.
* Progress on downloadAndProcess.py:
    - Each line in the delegated file corresponding to IPv4 resources
is converted into as many lines including CIDR blocks as necessary.
    - The resources (both IPv4 and IPv6) delegated to each organization
are aggregated as much as possible
    - Each aggregated delegated block is compared to routed blocks to
determine how it is being routed.
    - Once we have all the sub-blocks being routed for each aggregated
delegated block, the visibility and the deaggregation for the delegated
block are computed.
    - Finally, for each organization we generate a summary including
all the blocks that were delegated to it, all the blocks being routed,
the average visibility and the average deaggregation.
* The README file was updated.