Progress on BGPDataHandler
* Default value for the urls_file variable modified in order to use
BGPoutputs.txt which has a URL pointing to a ‘show ip bop’ output from
AS2.0
* PyTricia for prefixes modified in order to have as data a list of
indexes of the rows in the bgp_data DataFrame for the corresponding
prefix. This is to avoid having duplicate data in two different
structures.
* Methods added and updated in order to be able to process both RIB
files and ‘show ip bop’ outputs
* Added code necessary to merge the partial results of the processing
of different files in case the urls_file contains more than one URL.