Fixes on delegation stats computation and try to make delegated_stats script perform better
delegated_stats_v4.py was created. Instead of initializing the stats_df
DataFrame with all the possible combinations of the index columns
values, rows are created when necessary. I thought this would perform
better, however, the fact of having to sort the index after adding a
new row probably makes it perform worse. Indeed, comparing the
execution of v3 of the script and v4 of the script I found out that v3
is faster.