Progress on computation of routing stats
* BGPDataHandler updated in order to have different PyTricia for IPv4
and IPv6 and to include variables with the longest prefix seen in the
analyzed BGP data.
* Computation of routing statistics -> For each delegated block or
block resulting of aggregating delegated blocks, all the routed more
specific and less specific blocks are obtained. If there is at least
one less specific block being routed, the visibility is 100%. Else, the
visibility of the block is computed based on the coverage of the more
specific blocks.