Updates in computation of routing stats
* BGPDataHandler updated in order to include a dictionary with the
periods of time during which ASNs propagated prefixes.
* Added function in DelegatedHandler which returns a Radix with
information about all the delegations of IP prefixes available
* Function in routing_stats_prerfixesAndASes that computes statistics
for prefixes was completely restructured in order to fix classification
of prefixes and for better performance and clarity
* Added computation of average, standard deviation, max and min AS path
length for each routed prefix.
* Function that computes stats for ASes updated in order to consider
historical info not only for ASes that have been active as origin ASes
but also for ASes that have been active as middle ASes (propagating
prefixes).
* Function that computes stats for ASes updated in order to also
compute the number of upstreams for each delegated AS (using CAIDA’s
as-rel file)
* routing_stats_prefixesAndASes updated in order to save to a pickle
file a dictionary containing the visibility per periods of each
delegated prefix