Classification of prefixes based on comparison between origin Ases and AS paths updated
After talking to Geoff about the classification described in his paper
(SOSP, SODP, DOSP, DODP), we concluded that this classification is
useless and it’s only worth comparing the origin AS (same origin AS =
traffic engineering, different origin AS = hole punching).
The implementation of this classification was updated in order to
classify the prefixes in deaggregated (same origin AS) or delegated
(different origin AS). In any case the Levenshtein distance between the
AS paths is computed.