Dictionaries for ASes removed. Returning the BGP DataFrame instead. And some other changes.
This saves time of loading the dictionaries. It is not worth spending
time doing this as we only use these dictionaries to compute the number
of prefixes propagated and originated by each AS and this info can be
obtained from the DataFrame.
Some other changes for better performance and readability of the code.