Updates for better performance and readability and finished implementing necessary code to insert info about routing data into the DB
Instead of having redundant data in the routing files and in the DB, we
will only insert into the DB the path to the routing file for a given
routing date. This way the routing_data table in the DB is used as an
index to get the path to the right file to use for a given routing date.