Changes due to insertion of paths to routing files into DB
As we will have the table routing_data as an index to get the right
routing file for a given date, we had to update the class
BGPDataHandler in order to get the list of file paths for routing files
from the DB instead of walking the archive folder.
We decided to do this because the date in the name of the routing files
is not the real routing date, but it has an offset of one or two days.