New script to compute stats for old dates
Implemented script that will compute routing statistics and statistics
about update rates and probability of deaggregation for all the old
dates for which we have routing and updates data available.
The script will check whether a by-rib file is available first and if
there is, if the corresponding readable file is already present in the
readable folder (to save time). If no bgprib file is available, the
available dmp files will be used.