Updates related to partitions for updates table in DB
dailyExecution:
* If a new year has started, a new partition is created with the
corresponding indexes and constraints and the updates_insert_trigger
function is recreated adding a line for the new year.
* The copy command is executed on the right partition unless it is the
first day or the last day of the year in which case the CSV file may
contain BGP updates for a different year, therefore the insertion is
made into the master table (updates) letting the updates_insert_trigger
function to redirect the rows to the right partition.
generateSQLfilesForInsertionIntoDB:
* The copy command is executed on the right partition unless it is the
first day or the last day of the year in which case the CSV file may
contain BGP updates for a different year, therefore the insertion is
made into the master table (updates) letting the updates_insert_trigger
function to redirect the rows to the right partition.